Moved fileIcon to commons folder to re-use in the upload modal.
