+++ src/components/Common/FileIcon.jsx
@@ -2,11 +2,11 @@ import React from "react";
 
 import { FileIcon, defaultStyles } from "react-file-icon";
 
-const File = ({ fileName }) => {
+const File = ({ fileName, className }) => {
   const extension = fileName.match(/([^.]+)$/)[1];
 
   return (
-    <div className="ne-attachments__preview-wrapper__icon">
+    <div {...{ className }}>
       <FileIcon
         {...{ extension }}
         labelColor="#4558f9"
